# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: BX YunPan API
  version: 0.1.0
  description: |
    Contract-first API for the Go backend of BX YunPan.
    Large file bytes are transferred directly between the browser and S3-compatible storage.
  license:
    name: MIT
    url: https://opensource.org/license/mit
servers:
  - url: /
    description: Current origin behind the deployment proxy
...
security:
  - bearerAuth: []
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in; the refresh token is set as an HttpOnly cookie
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate a refresh session
      operationId: refreshSession
      security: []
      parameters:
        - $ref: '#/components/parameters/Origin'
      responses:
        '200':
          description: Access token refreshed and refresh cookie rotated
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current refresh session
      operationId: logout
      parameters:
        - $ref: '#/components/parameters/Origin'
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Get the current user
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/users/me/quota:
    get:
      tags: [Users]
      summary: Get storage quota
      operationId: getCurrentUserQuota
      responses:
        '200':
          description: Quota usage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quota'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /openapi.yaml:
    get:
      tags: [Operations]
      summary: Get the OpenAPI contract
      operationId: getOpenAPIContract
      security: []
      responses:
        '200':
          description: OpenAPI YAML document
          content:
            application/yaml:
              schema:
                type: string
        '403':
          $ref: '#/components/responses/Forbidden'
  /docs:
    get:
      tags: [Operations]
      summary: Get interactive API documentation
      operationId: getAPIDocs
      security: []
      responses:
        '200':
          description: Interactive API documentation
          content:
            text/html:
              schema:
                type: string
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    shareToken:
      type: apiKey
      in: header
      name: X-Share-Token
  parameters:
...
    Origin:
      name: Origin
      in: header
      required: true
      schema:
        type: string
        format: uri
  responses:
    Unauthorized:
      description: Authentication or token validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: The deployment access policy denied this request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Conflict:
      description: Resource state, version, name, or idempotency conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    TooManyRequests:
      description: Rate or quota limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    UUID:
      type: string
      format: uuid
      example: 0198a1f0-7b2a-7000-8000-000000000001
...
    Timestamp:
      type: string
      format: date-time
...
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: upload.hash_mismatch
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        request_id:
          type: string
...
    User:
      type: object
      required: [id, username, status, created_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        username:
          type: string
        email:
          type: string
          format: email
          nullable: true
        status:
          type: string
          enum: [active, disabled]
        created_at:
          $ref: '#/components/schemas/Timestamp'
    Quota:
      type: object
      required: [quota_bytes, used_logical_bytes, reserved_bytes]
      properties:
        quota_bytes:
          type: integer
          format: int64
          minimum: 0
        used_logical_bytes:
          type: integer
          format: int64
          minimum: 0
        reserved_bytes:
          type: integer
          format: int64
          minimum: 0
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 10
          maxLength: 128
      additionalProperties: false
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          maxLength: 254
        password:
          type: string
          format: password
          maxLength: 128
      additionalProperties: false
    AuthResponse:
      type: object
      required: [access_token, expires_in, user]
      properties:
        access_token:
          type: string
          writeOnly: true
        token_type:
          type: string
          enum: [Bearer]
          default: Bearer
        expires_in:
          type: integer
          minimum: 1
        user:
          $ref: '#/components/schemas/User'
...